requirements.txt
streamlit
google-genai
yt-dlp
python-dotenv
pydantic

